remove unnesessary files
