+++ src/zkevm_specs/__init__.py
@@ -1,9 +1,7 @@
 from . import bytecode
 from . import copy_circuit
-from . import encoding
 from . import evm
 from . import exp_circuit
-from . import opcode
 from . import state_circuit
 from . import tx
 from . import util
