--- src/zkevm_specs/encoding/__init__.py
@@ -1,2 +1,0 @@
-from .utils import *
-from .lookup import LookupTable
