--- src/zkevm_specs/opcode/__init__.py
@@ -1,1 +1,0 @@
-from .signextend import *
