+++ src/zkevm_specs/state_circuit.py
@@ -4,8 +4,7 @@ from math import log, ceil
 
 from zkevm_specs.evm.table import MPTProofType
 
-from .util import FQ, RLC, U160, U256, Expression, linear_combine_bytes
-from .encoding import U8, is_circuit_code
+from .util import FQ, RLC, U8, U160, U256, Expression, linear_combine_bytes, is_circuit_code
 from .evm import (
     RW,
     AccountFieldTag,
